New account with it group service class

+++ app/Services/BaseService.php
@@ -2,7 +2,9 @@
 
 namespace App\Services;
 
+use App\Exceptions\ServiceException;
 use App\Services\Interfaces\BasicOperation;
+use Illuminate\Support\Collection;
 
 class BaseService implements BasicOperation
 {
@@ -22,16 +24,22 @@ class BaseService implements BasicOperation
         return $this->_model;
     }
 
-    public function store(mixed $model = null, array $data = []): bool
+    public function store(mixed $model = null, Collection $data): bool
     {
-        $this->setModel($model->create($data));
+        if(is_null($model))
+            throw new ServiceException('Model Not Found');
+
+        $this->setModel($model->create($data->toArray()));
         
         return !is_null($this->getModel());
     }
     
-    public function update(mixed $model = null, array $data = []): bool
+    public function update(mixed $model = null, Collection $data): bool
     {
-        $is_updated = $model->update($data);
+        if(is_null($model))
+            throw new ServiceException('Model Not Found');
+
+        $is_updated = $model->update($data->toArray());
 
         if($is_updated)
             $this->setModel($model->fresh());
@@ -41,6 +49,9 @@ class BaseService implements BasicOperation
     
     public function destroy(mixed $model = null): bool
     {
+        if(is_null($model))
+            throw new ServiceException('Model Not Found');
+
         $is_destroyed = $model->delete();
 
         if($is_destroyed)
